Update and rename compare.py to lesson03_compare.py
This has some code in it that's useless right now because I misunderstood the question, but I wanted to keep it for future reference. The dict_to_csv function turns a dictionary into a csv but assumes the dictionary is ordered, which it's not.

This code will add anyone who is not found in survey.csv to a csv called "extra_employees.csv" whereas the prompt was to add the extra information about employees to their own rows in all_employees.csv.