Update lesson03_compare.py
There has to be a better  way to do this and I think I might still be relying on dictionaries being ordered (which they aren't) but this does seem to write the appended employee information to a new file called appended_employees.csv. I cheated (I think) by adding entries like twitter:"" and github:"" to the dictionary on which appended_employees is based. Will try to fix.